requirements.txt
SQLAlchemy~=2.0.36
python-gvm~=24.12.0
lxml~=5.3.0
psycopg2-binary~=2.9.10
pymysql~=1.1.1
python-dotenv~=1.0.1
typer~=0.15.1
rich~=13.7.1
InquirerPy~=0.3.4